+++ src/components/content/content.tsx
@@ -1,0 +1,9 @@
+import { Component, } from '@stencil/core'
+
+@Component({
+  tag: 'slate-content',
+  styleUrl: './content.css',
+})
+class Content {}
+
+export { Content }