+++ stencil.config.ts
@@ -1,17 +1,26 @@
-import { Config } from '@stencil/core';
+import { Config } from '@stencil/core'
 
-export const config: Config = {
-  namespace: 'slate-ui-stencil',
+const config: Config = {
+  namespace: 'slate-ui',
+  globalStyle: 'src/theme.css',
   outputTargets: [
     {
       type: 'dist',
       esmLoaderPath: '../loader',
-      copy: [{ src: 'theme.css' }, { src: 'theme-typography.css' }, { src: 'styles.css' }],
+      copy: [
+        { src: 'theme.css' },
+        { src: 'theme-typography.css' },
+        { src: 'styles.css' },
+      ],
     },
     {
       type: 'dist-custom-elements',
       customElementsExportBehavior: 'auto-define-custom-elements',
-      copy: [{ src: 'theme.css' }, { src: 'theme-typography.css' }, { src: 'styles.css' }],
+      copy: [
+        { src: 'theme.css' },
+        { src: 'theme-typography.css' },
+        { src: 'styles.css' },
+      ],
     },
     {
       type: 'docs-readme',
@@ -20,4 +29,6 @@ export const config: Config = {
   testing: {
     browserHeadless: 'new',
   },
-};
+}
+
+export { config }